services/etl/requirements.txt
pandas
nltk
kafka-python
